repository: fix left-over file for symlink check
To determine whether the host we're running on supports symlinks, we
try to create a temporary symlink in the repo's working directory. If
creation succeeded, then symlinks are supported. This works just fine on
non-Windows hosts, but on Windows hosts it does leave behind the test
file without deleting it correctly.

The reason for this is that on Windows, one has to delete symlinks to
files with `DeleteFile`, but symlinks to directories with
`RemoveDirectory`. As we are creating the symlink with a non-existing
target of "testing", the symlink will be a directory symlink, but we
still use `p_unlink` which will map to `DeleteFile` and thus deletion
fails. Note: in the case where the target repo has a file "testing"
inside of its working directory, then we'd actually succeed to unlink
the file.

The issue can be fixed by creating another file and link to that one.
One might think of easier alternatives, e.g. linking to ".git" and then
just using `p_rmdir` on Windows. But this would fall flat for e.g.
submodules, which have a ".git" file, and for cases where the workdir
and gitdir live in different places.

Furthermore it's debatable whether `p_unlink` should emulate unlink(3P)
more closely. But this would require us to check what kind of file we
want to unlink, which could potentially have a non-negligible impact on
performance.