cmake: add embedded libssh2 as subdirectory
Due to recent changes to some required defines in libssh2, we now fail
to build the library correctly. This results from the fact that we're
manually including libssh2 sources and its include directory into our
own build infrastructure, which may not be configured as required by
libssh2.

As libssh2 is driven by CMake as well, we can simply include it as a
subdirectory and stop caring for its configuration. Instead, we will
now just link against the generated library and be done with it, without
caring about the library internals.