stream: introduce `poll` callback
The current stream backends do not allow for setting up any timeouts,
which are required for a proper keepalive implementation. Thus, we now
add a new callback `poll` to the stream backend that accepts a timeout
value as parameter. The function will return as soon as there is any
data available for the stream to be read or, if no data was received in
the given time, it will return zero.

To not break existing external implementations, `git_stream_poll` will
return a positive value immediately when the backend doesn't provide a
polling function. While this effectively disallows keepalive to work
with such backends, this is no worse than it has been previously and we
will simply block until either any data has been received in a
subsequent read or until the TCP connection has been closed.